==> notebooks/_History/03_text_analysis_normalization_PCA.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.decomposition import PCA
from transformers import BertTokenizer, BertModel, pipeline
import torch
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
import ast
import re
from datetime import datetime

Import data

# Upload the data frame with BERT embeddings:
file_path_embeddings = r'.\..\data\embeddings_output\df_user_embeddings_BERT.csv' # Update this with your file path
df_user_embeddings = pd.read_csv(file_path_embeddings)

df_user_embeddings.head(2)

# Upload user DataFrame:
file_path_user_df = r'.\..\data\data_clean\user_clean_data_ecommerce.csv'  # Update this with your file path
df_user = pd.read_csv(file_path_user_df)
df_user.shape

missing_values = df_user.isnull().sum()
missing_values

df_user.dropna(inplace=True)
df_user.shape

### 1. Normalization of BERT embeddings:

# Select only columns with embedding vectors (e.g., `dim_0`, `dim_1`, ..., `dim_780`)
embedding_columns = [col for col in df_user_embeddings.columns if col.startswith('dim_')]
# Convert to NumPy array
embedding_matrix = df_user_embeddings[embedding_columns].values  
embedding_matrix

# Compute L2 norm along rows
l2_norms = np.linalg.norm(embedding_matrix, axis=1, keepdims=True)  

# Normalize the embedding columns
normalized_embeddings = embedding_matrix / l2_norms  # Divide each row by its L2 norm
normalized_embeddings

Generate the output DataFrame with normalized embeddings:

num_dimensions = normalized_embeddings.shape[1]
norm_embedding_columns_names = [f'dim_norm_{i}' for i in range(num_dimensions)]

embeddings_norm = pd.DataFrame(normalized_embeddings, columns=norm_embedding_columns_names)

df_normalized_embeddings = pd.concat([df_user.reset_index(drop=True), embeddings_norm], axis=1)
df_normalized_embeddings.head(2)

# Identify columns of interest (dim_0 to dim_767)
columns_of_interest = [f'dim_norm_{i}' for i in range(768)]

# Count rows with NaN in these columns
num_rows_with_nan_norm = df_normalized_embeddings[columns_of_interest].isna().any(axis=1).sum()

# Display the count
print(num_rows_with_nan_norm)

### 2. Dimentionality reduction with PCA

* Similar steps but here the output is a final user dataframe with vectors of 300 dimensions (In the beginning I have chosen 10 only because that's what they have done in the artile but the variance dropped to 45%, so I set up for now 300 which gives ~95%.) We can always adjust it

pd.set_option('display.max_colwidth', 50)  # Adjust column width
pd.set_option('display.width', 100)       # Adjust width of display

# Perform dimensionality reduction to 300 components using PCA
pca = PCA(n_components=300)

pca_embeddings = pca.fit_transform(normalized_embeddings)

print(pca_embeddings.shape)

pca_embeddings

### 3. PCA - Variance analysis

* Verify if the size of reduced embeddings is big enough to still keep the meaning of the textual data

embeddings_norm.shape

# Fit PCA to the original embeddings
pca = PCA(n_components=300)
pca.fit(embeddings_norm)

# Check explained variance
explained_variance = pca.explained_variance_ratio_
total_variance = sum(explained_variance)

print("Explained Variance Ratio for 300 Dimensions:", explained_variance)
print("Total Explained Variance Captured by 300 Dimensions:", total_variance)

### 4. Generate a merged DataFrame from df_user, including columns for embedding dimensions.

# Create column names for embedding dimensions - normalized data after dimentionality reducion
num_dimensions_pca = pca_embeddings.shape[1]
normalized_columns_PCA = [f'dim_norm_PCA{i}' for i in range(num_dimensions_pca)]

embeddings_df_PCA = pd.DataFrame(pca_embeddings, columns=normalized_columns_PCA)

df_user_normalized_PCA = pd.concat([df_user.reset_index(drop=True), embeddings_df_PCA], axis=1)

df_user_normalized_PCA.head(5)

df_user_normalized_PCA[df_user_normalized_PCA['user_ID'] == 'AEVTGJFLW22HVSWOJLJCBJUN46WA']

Save the output DataFrame:

# Save the DataFrame with normalized & reduced embeddings:
timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")

file_path = rf'.\..\data\embeddings_dim_reduction\df_user_normalized_PCA{timestamp}.csv'
df_user_normalized_PCA.to_csv(file_path, index=False)
